Splitting up integration tests for circleci
